Stacks and Queues problems are done
The two stack based queue was different. I disagreed with some of the discussions found. The solutions that put the heavy lifting on push would have required O(n) on each push which seemed like a waste. I decided to instead put these in batches on the output so that if the shift were frequent, these would be small, resulting in a much lower theta(n).