Find a duplicate and find a missing number in arrays
These are two very different solutions because of a few edge cases. In the missing case, find a sum total of all elements and getting the difference between that and what the total should be returns the result. FindDuplcate only allows for numbers between 1 and n which allows for a tortoise and hare solution to find a loop caused by the indexes of the array. These each could have also be solved with a for loop and a hashmap for a slightly slower runtime.