Two Solutions
One is about as good as it can get. It was adapted from some other solutions. I want to shift to getting closer to the top answerers on leetcode. The other passes everything except the stress test. It seems to be similar to the other recommended answers online. I'll come back to it later and either optimize it or entirely rewrite it.