Blind 75 String Problems done; Two problems solved for O(N^2)
I found a quick solution as O(N^2) on the longest palindromic substring. I decided to check some other solutions to see if there was a better solution but they were all O(N^2) as well. Palindromic Substrings was an easy solution after the first one. It is mostly an adapted solution.